fix: avoid unnecessary board flip, thanks to Larry Isaacs
"The reason that animations don't work with the board flipped is due to
the resize2 procedure in board.tcl. Its comment mentions "incoherent
state" if the board is resized while flipped, so in this case it
un-flips the board, calls the normal resize procedure, and then re-flips
the board.  This has the unfortunate side effect of canceling the
animation.  I don't know if the "incoherent state" still applies, but
the attached patch assumes that it does.

For animations, the board size isn't changing,  This means means that
the normal resize procedure will exit early since nothing needs to be
done.  In this case there doesn't seem to be much need to un-flip and
re-flip the board either.  The patch adds a check in resize2 to see if
the size is changing when the board is currently flipped and skips the
entire un-flip, resize, re-flip process if the size isn't changing.
This keeps the animation from being canceled."