Fix scorebar crash
Start with a small board with auto-resize enable.
Enlarging the width of the board's window will reach the point when the
scorebar's width will change from 10 to 14. But the board's window now
doesn't have enough space for the board and it is shrinked, but now the
scorebar's width will change againt from 14 to 10, etc....
The infinite loop in the end will cause a crash by the tcl library.