moveEntry_char: Fix the key bindings and move list sorting.
Bind all the "normal" (not Ctrl+ or Alt+) chars to moveEntry_char and
let the event propagate only if the char is not part of a valid move.
The moves matching the case of the inserted text must be at the beginning
of the list to deal with ambiguities like bxc4 vs Bxc4.